Movido logica a gameservice solucionados algunos errores de dobles subscribes
